protocol: align cross-language wire format and signing layout
Closes the wire-compatibility findings from the 2026-05-02 audit. After this
commit, every implementation produces byte-identical MeshPacket bytes for the
same input, and signature verification works across language boundaries.

Wire format
- C# PacketSerializer: write Guid in RFC 4122 big-endian (was .NET's mixed-
  endian default — broke RREQ dedup hashing the moment routing shipped).
- Swift / C: TTL widened from UInt8 / uint8_t to Int32 (was silently
  truncating values > 255).
- Go SosPriority retyped from byte (couldn't hold 999) to int32; canonical
  on-wire value is 255 anyway, but the constant must be representable.

Signing
- C# PacketSigningService.BuildSignableData rewritten: little-endian + 4-byte
  Type / Priority (was big-endian via Array.Reverse + 1-byte Type, which
  broke verification against every other language).
- Nonce dedup keyed by (source, nonce) tuple (was hex(nonce) only — collision
  surface across senders).

Other security primitives
- Python _zero_memory now actually zeroes (was a no-op despite the name).
- Kotlin Ed25519Service: imported java.security.SecureRandom (was
  org.bouncycastle.crypto.SecureRandom which doesn't exist).
- Rust main.rs disabled — its security/* dependencies don't compile and the
  demo is unrelated to the protocol library itself.

Constants
- SosPriority canonicalised to 255 across all 8 languages (was 999 in some).

+++ swift/Sources/AetherProtocol/Constants.swift
@@ -5,8 +5,9 @@ import Foundation
 /// Protocol-level constants matching the Aether specification.
 public struct ProtocolConstants {
     // MARK: - Routing
-    public static let defaultTtl: UInt8 = 7
-    public static let sosTtl: UInt8 = 15
+    // Wire format is Int32 LE; constants typed Int32 so they assign directly into MeshPacket.ttl.
+    public static let defaultTtl: Int32 = 7
+    public static let sosTtl: Int32 = 15
     public static let routeTimeoutMs: Int = 5000
     public static let routeExpirySeconds: Int = 300
